webapi/tct-batterystatus-w3c-tests/batterystatus/w3c/battery-plugging-in-dischargingTime-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Battery Test: battery-plugging-in-dischargingTime</title>
<meta name="flags" content="interact">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #note {
    background-color: #fef1b5;
    border: solid 1px #cdab2d;
    padding: 5px;
    margin: 15px;
    display: block;
  }
</style>

<h2>Description</h2>
<p>
  This test validates that ondischargingtimechange and dischargingTime attributes exist and
  are set to correct values, with corresponding events fired,
  when the charger is plugged in.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    The device is unplugged from the charger before this test is run.
  </li>
  <li>
    The battery must not be full or reach full capacity before the time the test is run.
  </li>
</ol>

<div id="note">
  <ol>
    <li>
      Plug in the charger and wait for the test to complete.
    </li>
    <li>
      The tests may take long time since the definition of how
      often the dischargingtimechange event is fired is left to the implementation.
    </li>
  </ol>
</div>

<div id="log"></div>

<script>

(function() {

  setup({ explicit_timeout: true });

  var ondischargingtimechange_test = async_test(
      'When the device is plugged in and its discharging time is updated, ' +
      'must set the dischargingTime attribute\'s value to Infinity and ' +
      'fire a dischargingtimechange event.');

  var batterySuccess = function (battery) {
    battery.ondischargingtimechange = ondischargingtimechange_test.step_func(function () {
      if (battery.charging) {
        assert_equals(battery.dischargingTime, Infinity,
            'The value of the dischargingTime attribute must be set to Infinity.');
        ondischargingtimechange_test.done();
      }
    });
  };

  var batteryFailure = function (error) {
    ondischargingtimechange_test.step(function () {
      assert_unreached(error.message);
    });
    ondischargingtimechange_test.done();
  };

  navigator.getBattery().then(batterySuccess, batteryFailure);

})();

</script>
